pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swarm-agent-stack"
version = "0.1.0"
description = "Multi-agent orchestration framework"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "filelock>=3.10.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
]

[project.optional-dependencies]
memory = ["chromadb>=0.4.0"]
chain = ["web3>=6.0.0", "eth-account>=0.9.0"]
channels = ["python-telegram-bot>=21.0", "discord.py>=2.3", "lark-oapi>=1.3.0"]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3"]
feishu = ["lark-oapi>=1.3.0"]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0"]

[project.scripts]
swarm = "main:main"

[tool.setuptools]
py-modules = ["main"]
packages = ["core", "adapters", "adapters.llm", "adapters.memory", "adapters.chain", "adapters.channels", "reputation"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
